refactor(tile): mem.rs was five files in one
3473 lines under one module doc that had to say "its layouts, windows, and the
memory-side Tile operations (construction, views, the cooperative copy)" —
a list, because there was no single subject. Now `tile/mem/`, one subject each:

- base.rs (282)    what a MemData is: Backing, Store, Access and the flags
- of.rs (564)      building one from a tensor, a fused sink or a fused source
- stage.rs (774)   deriving a shared-memory stage: StageForm and the smem ctors
- access.rs (1591) filling, the views a leaf reads through, and `at`
- window.rs (304)  GmemLayout, Window, SourceWindow

The free functions went to whichever section calls them, which the call graph
settled with no judgement calls: every one of the eighteen has exactly one
caller section, bar `stage_compaction`.

Splitting cost four items their privacy — `MemData::window_start`, `Padding`'s
three fields and `stage_compaction` are now `pub(crate)`. That is the coupling
the single file was hiding, not new coupling.

No behaviour change.
